// lib/types/user.ts
export interface Address {
	street: string
	suite: string
	city: string
	zipcode: string
	geo: {
		lat: string
		lng: string
	}
}

export interface Company {
	name: string
	catchPhrase: string
	bs: string
}

export interface User {
	id: number
	name: string
	username: string
	email: string
	address: Address
	phone: string
	website: string
	company: Company
}

// For table display - flattened structure
export interface UserTableData {
	id: number
	name: string
	email: string
	companyName: string
	city: string
}